chore: move tests from conftest
ECALC-1813
